+++ Thirdweb.Tests/Thirdweb.SmartWallet.Tests.cs
@@ -1,6 +1,5 @@
 using System.Numerics;
 using Nethereum.Hex.HexTypes;
-using Nethereum.RPC.Eth.DTOs;
 
 namespace Thirdweb.Tests;
