+++ Thirdweb.Tests/Thirdweb.Wallets.Tests.cs
@@ -1,5 +1,4 @@
 ﻿using Nethereum.Hex.HexTypes;
-using Nethereum.RPC.Eth.DTOs;
 
 namespace Thirdweb.Tests;
