+++ Thirdweb/Thirdweb.Extensions/ThirdwebExtensions.cs
@@ -3,9 +3,31 @@ using Nethereum.Hex.HexTypes;
 
 namespace Thirdweb
 {
-    public static class ThirdwebContractExtensions
+    public static class ThirdwebExtensions
     {
         #region Common
+
+        public static async Task<ContractMetadata> GetMetadata(this ThirdwebContract contract)
+        {
+            if (contract == null)
+            {
+                throw new ArgumentNullException(nameof(contract));
+            }
+
+            var contractUri = await ThirdwebContract.Read<string>(contract, "contractURI");
+
+            return await ThirdwebStorage.Download<ContractMetadata>(contract.Client, contractUri);
+        }
+
+        public static async Task<string> GetPrimarySaleRecipient(this ThirdwebContract contract)
+        {
+            if (contract == null)
+            {
+                throw new ArgumentNullException(nameof(contract));
+            }
+
+            return await ThirdwebContract.Read<string>(contract, "primarySaleRecipient");
+        }
 
         public static async Task<BigInteger> GetBalance(this ThirdwebContract contract)
         {
@@ -18,6 +40,83 @@ namespace Thirdweb
             var rpc = ThirdwebRPC.GetRpcInstance(client, contract.Chain);
             var balanceHex = await rpc.SendRequestAsync<string>("eth_getBalance", contract.Address, "latest").ConfigureAwait(false);
             return new HexBigInteger(balanceHex).Value;
+        }
+
+        public static async Task<byte[]> GetNFTImageBytes(this NFT nft, ThirdwebClient client)
+        {
+            if (client == null)
+            {
+                throw new ArgumentNullException(nameof(client));
+            }
+
+            return string.IsNullOrEmpty(nft.Metadata.Image) ? new byte[] { } : await ThirdwebStorage.Download<byte[]>(client, nft.Metadata.Image).ConfigureAwait(false);
+        }
+
+        public static async Task<BigInteger> GetBalance(this IThirdwebWallet wallet, ThirdwebClient client, BigInteger chainId, string erc20ContractAddress = null)
+        {
+            if (wallet == null)
+            {
+                throw new ArgumentNullException(nameof(wallet));
+            }
+
+            if (client == null)
+            {
+                throw new ArgumentNullException(nameof(client));
+            }
+
+            if (chainId <= 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(chainId), "Chain ID must be greater than 0.");
+            }
+
+            var address = await wallet.GetAddress().ConfigureAwait(false);
+
+            if (erc20ContractAddress != null)
+            {
+                var erc20Contract = await ThirdwebContract.Create(client, erc20ContractAddress, chainId).ConfigureAwait(false);
+                return await erc20Contract.ERC20_BalanceOf(address).ConfigureAwait(false);
+            }
+
+            var rpc = ThirdwebRPC.GetRpcInstance(client, chainId);
+            var balanceHex = await rpc.SendRequestAsync<string>("eth_getBalance", address, "latest").ConfigureAwait(false);
+            return new HexBigInteger(balanceHex).Value;
+        }
+
+        public static async Task<ThirdwebTransactionReceipt> Transfer(this IThirdwebWallet wallet, ThirdwebClient client, BigInteger chainId, string toAddress, BigInteger weiAmount)
+        {
+            if (wallet == null)
+            {
+                throw new ArgumentNullException(nameof(wallet));
+            }
+
+            if (client == null)
+            {
+                throw new ArgumentNullException(nameof(client));
+            }
+
+            if (chainId <= 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(chainId), "Chain ID must be greater than 0.");
+            }
+
+            if (string.IsNullOrEmpty(toAddress))
+            {
+                throw new ArgumentException(nameof(toAddress), "Recipient address cannot be null or empty.");
+            }
+
+            if (weiAmount < 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(weiAmount), "Amount must be 0 or greater.");
+            }
+
+            var txInput = new ThirdwebTransactionInput()
+            {
+                From = await wallet.GetAddress().ConfigureAwait(false),
+                To = toAddress,
+                Value = new HexBigInteger(weiAmount)
+            };
+            var tx = await ThirdwebTransaction.Create(client, wallet, txInput, chainId).ConfigureAwait(false);
+            return await ThirdwebTransaction.SendAndWaitForTransactionReceipt(tx).ConfigureAwait(false);
         }
 
         #endregion
@@ -590,6 +689,371 @@ namespace Thirdweb
             {
                 return await ThirdwebContract.Read<BigInteger>(contract, "totalSupply");
             }
+        }
+
+        #endregion
+
+        #region NFT - ERC721
+
+        public static async Task<NFT> ERC721_GetNFT(this ThirdwebContract contract, BigInteger tokenId)
+        {
+            if (contract == null)
+            {
+                throw new ArgumentNullException(nameof(contract));
+            }
+
+            var uri = await contract.ERC721_TokenURI(tokenId).ConfigureAwait(false);
+            var metadata = await ThirdwebStorage.Download<NFTMetadata>(contract.Client, uri).ConfigureAwait(false);
+            metadata.Id = tokenId.ToString();
+
+            var owner = Constants.ADDRESS_ZERO;
+            try
+            {
+                owner = await contract.ERC721_OwnerOf(tokenId).ConfigureAwait(false);
+            }
+            catch (Exception)
+            {
+                owner = Constants.ADDRESS_ZERO;
+            }
+
+            return new NFT
+            {
+                Metadata = metadata,
+                Owner = owner,
+                Type = NFTType.ERC721,
+                Supply = 1,
+            };
+        }
+
+        public static async Task<List<NFT>> ERC721_GetAllNFTs(this ThirdwebContract contract, BigInteger? startTokenIdIncluded = null, BigInteger? endTokenIdExcluded = null)
+        {
+            if (contract == null)
+            {
+                throw new ArgumentNullException(nameof(contract));
+            }
+
+            if (startTokenIdIncluded == null)
+            {
+                startTokenIdIncluded = 0;
+            }
+
+            if (endTokenIdExcluded == null)
+            {
+                var totalSupply = await contract.ERC721_TotalSupply().ConfigureAwait(false);
+                endTokenIdExcluded = totalSupply;
+            }
+
+            var nftTasks = new List<Task<NFT>>();
+            for (var i = startTokenIdIncluded.Value; i < endTokenIdExcluded.Value; i++)
+            {
+                nftTasks.Add(contract.ERC721_GetNFT(i));
+            }
+
+            var allNfts = await Task.WhenAll(nftTasks).ConfigureAwait(false);
+            return allNfts.ToList();
+        }
+
+        public static async Task<List<NFT>> ERC721_GetOwnedNFTs(this ThirdwebContract contract, string owner)
+        {
+            if (contract == null)
+            {
+                throw new ArgumentNullException(nameof(contract));
+            }
+
+            if (string.IsNullOrEmpty(owner))
+            {
+                throw new ArgumentException("Owner must be provided");
+            }
+
+            try
+            {
+                var balanceOfOwner = await contract.ERC721_BalanceOf(owner).ConfigureAwait(false);
+                var ownedNftTasks = new List<Task<NFT>>();
+                for (var i = 0; i < balanceOfOwner; i++)
+                {
+                    var tokenOfOwnerByIndex = await contract.ERC721_TokenOfOwnerByIndex(owner, i).ConfigureAwait(false);
+                    ownedNftTasks.Add(contract.ERC721_GetNFT(tokenOfOwnerByIndex));
+                }
+                var ownedNfts = await Task.WhenAll(ownedNftTasks).ConfigureAwait(false);
+                return ownedNfts.ToList();
+            }
+            catch (Exception)
+            {
+                var allNfts = await contract.ERC721_GetAllNFTs().ConfigureAwait(false);
+                var ownedNfts = new List<NFT>();
+                foreach (var nft in allNfts)
+                {
+                    if (nft.Owner == owner)
+                    {
+                        ownedNfts.Add(nft);
+                    }
+                }
+                return ownedNfts.ToList();
+            }
+        }
+
+        #endregion
+
+        #region NFT - ERC1155
+
+        public static async Task<NFT> ERC1155_GetNFT(this ThirdwebContract contract, BigInteger tokenId)
+        {
+            if (contract == null)
+            {
+                throw new ArgumentNullException(nameof(contract));
+            }
+
+            var uri = await contract.ERC1155_URI(tokenId).ConfigureAwait(false);
+            var metadata = await ThirdwebStorage.Download<NFTMetadata>(contract.Client, uri).ConfigureAwait(false);
+            metadata.Id = tokenId.ToString();
+            var owner = string.Empty;
+            var supply = BigInteger.MinusOne;
+            try
+            {
+                supply = await contract.ERC1155_TotalSupply(tokenId).ConfigureAwait(false);
+            }
+            catch (Exception)
+            {
+                supply = BigInteger.MinusOne;
+            }
+
+            return new NFT
+            {
+                Metadata = metadata,
+                Owner = owner,
+                Type = NFTType.ERC1155,
+                Supply = supply,
+            };
+        }
+
+        public static async Task<List<NFT>> ERC1155_GetAllNFTs(this ThirdwebContract contract, BigInteger? startTokenIdIncluded = null, BigInteger? endTokenIdExcluded = null)
+        {
+            if (contract == null)
+            {
+                throw new ArgumentNullException(nameof(contract));
+            }
+
+            if (startTokenIdIncluded == null)
+            {
+                startTokenIdIncluded = 0;
+            }
+
+            if (endTokenIdExcluded == null)
+            {
+                var totalSupply = await contract.ERC1155_TotalSupply().ConfigureAwait(false);
+                endTokenIdExcluded = totalSupply;
+            }
+
+            var nftTasks = new List<Task<NFT>>();
+            for (var i = startTokenIdIncluded.Value; i < endTokenIdExcluded.Value; i++)
+            {
+                nftTasks.Add(contract.ERC1155_GetNFT(i));
+            }
+
+            var allNfts = await Task.WhenAll(nftTasks).ConfigureAwait(false);
+            return allNfts.ToList();
+        }
+
+        public static async Task<List<NFT>> ERC1155_GetOwnedNFTs(this ThirdwebContract contract, string owner)
+        {
+            if (contract == null)
+            {
+                throw new ArgumentNullException(nameof(contract));
+            }
+
+            if (string.IsNullOrEmpty(owner))
+            {
+                throw new ArgumentException("Owner must be provided");
+            }
+
+            var totalSupply = await contract.ERC1155_TotalSupply().ConfigureAwait(false);
+
+            var nfts = new List<NFT>();
+            for (var i = 0; i < totalSupply; i++)
+            {
+                var balanceOfOwner = await contract.ERC1155_BalanceOf(owner, i).ConfigureAwait(false);
+                if (balanceOfOwner > 0)
+                {
+                    var nft = await contract.ERC1155_GetNFT(i).ConfigureAwait(false);
+                    nft.QuantityOwned = balanceOfOwner;
+                    nfts.Add(nft);
+                }
+            }
+
+            return nfts;
+        }
+
+        #endregion
+
+        #region DropERC20
+
+        public static async Task<ThirdwebTransactionReceipt> DropERC20_Claim(this ThirdwebContract contract, IThirdwebWallet wallet, string receiverAddress, string amount)
+        {
+            if (contract == null)
+            {
+                throw new ArgumentNullException(nameof(contract));
+            }
+
+            if (wallet == null)
+            {
+                throw new ArgumentNullException(nameof(wallet));
+            }
+
+            if (string.IsNullOrEmpty(receiverAddress))
+            {
+                throw new ArgumentException("Receiver address must be provided");
+            }
+
+            if (string.IsNullOrEmpty(amount))
+            {
+                throw new ArgumentException("Amount must be provided");
+            }
+
+            var activeClaimCondition = await contract.GetActiveClaimCondition();
+
+            var erc20Decimals = await contract.ERC20_Decimals();
+
+            var rawAmountToClaim = BigInteger.Parse(amount.ToWei()).AdjustDecimals(18, erc20Decimals);
+
+            var isNativeToken = activeClaimCondition.Currency == Constants.NATIVE_TOKEN_ADDRESS;
+
+            var payableAmount = isNativeToken ? rawAmountToClaim * activeClaimCondition.PricePerToken : BigInteger.Zero;
+
+            // TODO: Merkle
+            var allowlistProof = new object[] { new byte[] { }, BigInteger.Zero, BigInteger.Zero, Constants.ADDRESS_ZERO };
+
+            var fnArgs = new object[]
+            {
+                receiverAddress, // receiver
+                rawAmountToClaim, // quantity
+                activeClaimCondition.Currency, // currency
+                activeClaimCondition.PricePerToken, // pricePerToken
+                allowlistProof, // allowlistProof
+                new byte[] { } // data
+            };
+
+            return await ThirdwebContract.Write(wallet, contract, "claim", payableAmount, fnArgs);
+        }
+
+        public static async Task<BigInteger> DropERC20_GetActiveClaimConditionId(this ThirdwebContract contract)
+        {
+            return await ThirdwebContract.Read<BigInteger>(contract, "getActiveClaimConditionId");
+        }
+
+        public static async Task<DropERC20_ClaimCondition> DropERC20_GetClaimConditionById(this ThirdwebContract contract, BigInteger claimConditionId)
+        {
+            return await ThirdwebContract.Read<DropERC20_ClaimCondition>(contract, "getClaimConditionById", claimConditionId);
+        }
+
+        public static async Task<DropERC20_ClaimCondition> GetActiveClaimCondition(this ThirdwebContract contract)
+        {
+            var activeClaimConditionId = await contract.DropERC20_GetActiveClaimConditionId();
+            return await contract.DropERC20_GetClaimConditionById(activeClaimConditionId);
+        }
+
+        #endregion
+
+        #region TokenERC20
+
+        public static async Task<ThirdwebTransactionReceipt> TokenERC20_MintTo(this ThirdwebContract contract, IThirdwebWallet wallet, string receiverAddress, string amount)
+        {
+            if (contract == null)
+            {
+                throw new ArgumentNullException(nameof(contract));
+            }
+
+            if (wallet == null)
+            {
+                throw new ArgumentNullException(nameof(wallet));
+            }
+
+            if (string.IsNullOrEmpty(receiverAddress))
+            {
+                throw new ArgumentException("Receiver address must be provided");
+            }
+
+            if (string.IsNullOrEmpty(amount))
+            {
+                throw new ArgumentException("Amount must be provided");
+            }
+
+            var erc20Decimals = await contract.ERC20_Decimals();
+
+            var rawAmountToMint = BigInteger.Parse(amount.ToWei()).AdjustDecimals(18, erc20Decimals);
+
+            return await ThirdwebContract.Write(wallet, contract, "mintTo", 0, receiverAddress, rawAmountToMint);
+        }
+
+        public static async Task<ThirdwebTransactionReceipt> TokenERC20_MintWithSignature(this ThirdwebContract contract, IThirdwebWallet wallet, TokenERC20_MintRequest mintRequest, string signature)
+        {
+            if (contract == null)
+            {
+                throw new ArgumentNullException(nameof(contract));
+            }
+
+            if (wallet == null)
+            {
+                throw new ArgumentNullException(nameof(wallet));
+            }
+
+            if (mintRequest == null)
+            {
+                throw new ArgumentNullException(nameof(mintRequest));
+            }
+
+            if (string.IsNullOrEmpty(signature))
+            {
+                throw new ArgumentException("Signature must be provided");
+            }
+
+            mintRequest = new TokenERC20_MintRequest
+            {
+                To = mintRequest.To ?? throw new ArgumentNullException(nameof(mintRequest.To)),
+                PrimarySaleRecipient = mintRequest.PrimarySaleRecipient ?? await contract.GetPrimarySaleRecipient(),
+                Quantity = mintRequest.Quantity > 0 ? mintRequest.Quantity : 1,
+                Price = mintRequest.Price,
+                Currency = mintRequest.Currency ?? Constants.NATIVE_TOKEN_ADDRESS,
+                ValidityStartTimestamp = mintRequest.ValidityStartTimestamp,
+                ValidityEndTimestamp = mintRequest.ValidityEndTimestamp > 0 ? mintRequest.ValidityEndTimestamp : Utils.GetUnixTimeStampIn10Years(),
+                Uid = mintRequest.Uid ?? Guid.NewGuid().ToByteArray()
+            };
+
+            var isNativeToken = mintRequest.Currency == Constants.NATIVE_TOKEN_ADDRESS;
+
+            var payableAmount = isNativeToken ? mintRequest.Quantity * mintRequest.Price : BigInteger.Zero;
+
+            return await ThirdwebContract.Write(wallet, contract, "mintWithSignature", payableAmount, mintRequest, signature);
+        }
+
+        public static async Task<(TokenERC20_MintRequest, string)> TokenERC20_GenerateMintSignature(this ThirdwebContract contract, IThirdwebWallet wallet, TokenERC20_MintRequest mintRequest)
+        {
+            if (contract == null)
+            {
+                throw new ArgumentNullException(nameof(contract));
+            }
+
+            if (wallet == null)
+            {
+                throw new ArgumentNullException(nameof(wallet));
+            }
+
+            if (mintRequest == null)
+            {
+                throw new ArgumentNullException(nameof(mintRequest));
+            }
+
+            var contractMetadata = await contract.GetMetadata();
+
+            var signature = await EIP712.GenerateSignature_TokenERC20(
+                domainName: contractMetadata.Name,
+                version: "1",
+                chainId: contract.Chain,
+                verifyingContract: contract.Address,
+                mintRequest: mintRequest,
+                signer: wallet
+            );
+
+            return (mintRequest, signature);
         }
 
         #endregion
