+++ Thirdweb/Thirdweb.Pay/ThirdwebPay.BuyWithCrypto.cs
@@ -1,5 +1,4 @@
 using System.Numerics;
-using System.Threading.Tasks;
 using Nethereum.Hex.HexTypes;
 
 namespace Thirdweb.Pay
