+++ Thirdweb/Thirdweb.Pay/ThirdwebPay.BuyWithFiat.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Threading.Tasks;
-
 namespace Thirdweb.Pay
 {
     public partial class ThirdwebPay
