+++ Thirdweb/Thirdweb.Wallets/InAppWallet/EmbeddedWallet/EmbeddedWallet.OAuth.cs
@@ -1,5 +1,3 @@
-using System.Threading.Tasks;
-using Nethereum.Web3.Accounts;
 using Newtonsoft.Json;
 
 namespace Thirdweb.EWS
